// Frontend/ucrs/src/types/user/user.ts
export interface UserData {
    id: number;
    email: string;
    first_name: string;
    middle_name: string;
    last_name: string;
    suffix: string;
    age: number;
    birthdate: string;
    gender: string;
    phone_number: string;
    barangay: string;
    full_address: string;
    zip_code: string;
    id_type: string;
    id_number: string;
    latitude?: string;
    longitude?: string;
    front_id: string;
    back_id: string;
    selfie_with_id: string;
    role: string;
}